+++ mw/extensions/DynamicPageList3/includes/HookHandlers/Installer.php
@@ -4,8 +4,8 @@ declare( strict_types = 1 );
 
 namespace MediaWiki\Extension\DynamicPageList4\HookHandlers;
 
-use MediaWiki\Extension\DynamicPageList4\Maintenance\CreateView;
 use MediaWiki\Extension\DynamicPageList4\Maintenance\DeleteTemplate;
+use MediaWiki\Extension\DynamicPageList4\Maintenance\DropView;
 use MediaWiki\Installer\Hook\LoadExtensionSchemaUpdatesHook;
 
 class Installer implements LoadExtensionSchemaUpdatesHook {
@@ -15,7 +15,7 @@ class Installer implements LoadExtensionSchemaUpdatesHook {
 	 * @codeCoverageIgnore Tested by updating or installing MediaWiki.
 	 */
 	public function onLoadExtensionSchemaUpdates( $updater ) {
-		$updater->addPostDatabaseUpdateMaintenance( CreateView::class );
 		$updater->addPostDatabaseUpdateMaintenance( DeleteTemplate::class );
+		$updater->addPostDatabaseUpdateMaintenance( DropView::class );
 	}
 }