+++ mw/extensions/DynamicPageList3/includes/Parse.php
@@ -15,6 +15,7 @@ use MediaWiki\Html\Html;
 use MediaWiki\MediaWikiServices;
 use MediaWiki\Parser\Parser;
 use MediaWiki\Parser\ParserOutputLinkTypes;
+use MediaWiki\Parser\PPFrame;
 use MediaWiki\Registration\ExtensionRegistry;
 use MediaWiki\Request\WebRequest;
 use MediaWiki\Title\Title;
@@ -90,7 +91,8 @@ class Parse {
 		Parser $parser,
 		array &$reset,
 		array &$eliminate,
-		bool $isParserTag
+		bool $isParserTag,
+		?PPFrame $frame
 	): string {
 		$dplStartTime = microtime( true );
 
@@ -99,8 +101,7 @@ class Parse {
 		$title = Title::castFromPageReference( $parser->getPage() );
 
 		// Check that we are not in an infinite transclusion loop
-		// @phan-suppress-next-line PhanDeprecatedProperty
-		if ( isset( $parser->mTemplatePath[$title->getPrefixedText()] ) ) {
+		if ( $frame && $title && !$frame->loopCheck( $title ) ) {
 			$this->logger->addMessage( Constants::WARN_TRANSCLUSIONLOOP, $title->getPrefixedText() );
 			return $this->getFullOutput( totalResults: 0, skipHeaderFooter: true );
 		}
@@ -146,10 +147,6 @@ class Parse {
 		}
 
 		$cleanParameters = Parameters::sortByPriority( $cleanParameters );
-
-		// To check if pseudo-category of Uncategorized pages is included
-		$this->parameters->setParameter( 'includeuncat', false );
-
 		foreach ( $cleanParameters as $parameter => $options ) {
 			foreach ( $options as $option ) {
 				// Parameter functions return true or false. The full parameter data will be
@@ -371,7 +368,7 @@ class Parse {
 
 			if ( $this->parameters->getParameter( 'goal' ) === 'categories' ) {
 				$pageNamespace = NS_CATEGORY;
-				$pageTitle = $row->cl_to;
+				$pageTitle = $row->lt_title;
 			} elseif ( $this->parameters->getParameter( 'openreferences' ) ) {
 				$imageContainer = $this->parameters->getParameter( 'imagecontainer' ) ?? [];
 				if ( $imageContainer !== [] ) {