+++ mw/extensions/DynamicPageList3/tests/phpunit/DPLQueryIntegrationTest.php
@@ -23,6 +23,17 @@ class DPLQueryIntegrationTest extends DPLIntegrationTestCase {
 		);
 	}
 
+	public function testFindPagesInNoneCategory(): void {
+		$this->assertArrayEquals(
+			[ 'DPLTestArticleNoCategory', 'DPLUncategorizedPage', 'DPLTestOpenReferences' ],
+			$this->getDPLQueryResults( [
+				'namespace' => NS_MAIN,
+				'category' => '_none_',
+			], '%PAGE%' ),
+			true
+		);
+	}
+
 	public function testFindPagesInCategoryWithOrderAndLimit(): void {
 		$this->assertArrayEquals(
 			[ 'DPLTestArticleMultipleCategories', 'DPLTestArticle 3' ],